+++ libs/designer/src/lib/core/actions/bjsworkflow/connections.ts
@@ -61,6 +61,8 @@ export const updateTemplateConnection = createAsyncThunk(
   'updateTemplateConnection',
   async (payload: ConnectionPayload, { dispatch, getState }): Promise<void> => {
     const { nodeId, connector, connection, connectionProperties, authentication } = payload;
+    const workflows = (getState() as TemplateRootState).template.workflows;
+    const defaultWorkflow = Object.values(workflows).length > 0 ? Object.values(workflows)[0] : undefined;
     dispatch(
       changeTemplateConnectionMapping({
         nodeId,
@@ -68,7 +70,7 @@ export const updateTemplateConnection = createAsyncThunk(
         connectionId: connection.id,
         authentication: authentication ?? getApiHubAuthenticationIfRequired(),
         connectionProperties: connectionProperties ?? getConnectionPropertiesIfRequired(connection, connector),
-        connectionRuntimeUrl: isOpenApiSchemaVersion((getState() as TemplateRootState).template.workflowDefinition)
+        connectionRuntimeUrl: isOpenApiSchemaVersion(defaultWorkflow?.workflowDefinition)
           ? connection.properties.connectionRuntimeUrl
           : undefined,
       })