+++ libs/designer/src/lib/ui/panel/templatePanel/quickViewPanel/tabs/workflowTab.tsx
@@ -1,5 +1,4 @@
-import type { AppDispatch, RootState } from '../../../../../core/state/templates/store';
-import { useSelector } from 'react-redux';
+import type { AppDispatch } from '../../../../../core/state/templates/store';
 import type { IntlShape } from 'react-intl';
 import constants from '../../../../../common/constants';
 import { useTheme } from '@fluentui/react';
@@ -8,9 +7,10 @@ import type { TemplatePanelTab } from '@microsoft/designer-ui';
 import { closePanel, openCreateWorkflowPanelView } from '../../../../../core/state/templates/panelSlice';
 import { clearTemplateDetails } from '../../../../../core/state/templates/templateSlice';
 import { LogEntryLevel, LoggerService, type Template } from '@microsoft/logic-apps-shared';
+import { useDefaultWorkflowTemplate } from '../../../../../core/state/templates/templateselectors';
 
 export const WorkflowPanel: React.FC = () => {
-  const { manifest, images } = useSelector((state: RootState) => state.template);
+  const { manifest, images } = useDefaultWorkflowTemplate() ?? {};
   const { isInverted } = useTheme();
   const imageName = useMemo(() => (isInverted ? images?.dark : images?.light), [isInverted, images]);