+++ libs/designer/src/lib/ui/templates/cards/templateCard.tsx
@@ -1,5 +1,5 @@
 import type { AppDispatch, RootState } from '../../../core/state/templates/store';
-import { changeCurrentTemplateName, loadTemplate } from '../../../core/state/templates/templateSlice';
+import { changeCurrentTemplateName } from '../../../core/state/templates/templateSlice';
 import { useDispatch, useSelector } from 'react-redux';
 import { Text } from '@fluentui/react-components';
 import { openQuickViewPanelView } from '../../../core/state/templates/panelSlice';
@@ -13,6 +13,7 @@ import type { OperationInfo } from '@microsoft/logic-apps-shared';
 import { equals, getBuiltInOperationInfo, isBuiltInOperation, LogEntryLevel, LoggerService } from '@microsoft/logic-apps-shared';
 import MicrosoftIcon from '../../../common/images/templates/microsoft.svg';
 import { PeopleCommunity16Regular } from '@fluentui/react-icons';
+import { loadTemplate } from '../../../core/actions/bjsworkflow/templates';
 
 interface TemplateCardProps {
   templateName: string;