+++ libs/designer/src/lib/ui/templates/connections/createConnection.tsx
@@ -2,13 +2,13 @@ import { useCallback, useMemo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import type { Connection } from '@microsoft/logic-apps-shared';
 import { equals } from '@microsoft/logic-apps-shared';
-import type { CreatedConnectionPayload } from '../connectionsPanel/createConnection/createConnectionWrapper';
-import { CreateConnectionInternal } from '../connectionsPanel/createConnection/createConnectionWrapper';
-import { useConnectorOnly } from '../../../core/state/connection/connectionSelector';
+import type { CreatedConnectionPayload } from '../../panel/connectionsPanel/createConnection/createConnectionWrapper';
+import { CreateConnectionInternal } from '../../panel/connectionsPanel/createConnection/createConnectionWrapper';
 import type { AppDispatch, RootState } from '../../../core/state/templates/store';
 import { getAssistedConnectionProps } from '../../../core/utils/connectors/connections';
 import { updateTemplateConnection } from '../../../core/actions/bjsworkflow/connections';
 import { useIntl } from 'react-intl';
+import { useConnector } from '../../../core/state/connection/connectionSelector';
 
 export const CreateConnectionInTemplate = (props: {
   connectorId: string;
@@ -19,11 +19,11 @@ export const CreateConnectionInTemplate = (props: {
   const intl = useIntl();
   const { connectorId, connectionKey, onConnectionCreated, onConnectionCancelled } = props;
   const dispatch = useDispatch<AppDispatch>();
-  const { data: connector } = useConnectorOnly(connectorId);
+  const { data: connector } = useConnector(connectorId);
 
   const { references, connections } = useSelector((state: RootState) => ({
     references: Object.keys(state.workflow.connections.references),
-    connections: state.template.manifest?.connections,
+    connections: state.template.connections,
   }));
   const isInAppConnector = equals(connections?.[connectionKey]?.kind, 'inapp');